more changes and separation of CigarAlignment class
